// back-office/src/app/openapi/model/currencyViewDTO.ts
/**
 * FinalStake
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CurrencyViewDTO { 
    code?: string;
    buyingConversion?: number;
    sellingConversion?: number;
    id?: number;
    name?: string;
}